+++ apps/dialtone-documentation/docs/.vuepress/baseComponents/UtilityClassTable.vue
@@ -1,17 +1,25 @@
 <template>
-  <table class="d-table dialtone-doc-table d-fc-primary">
-    <thead>
-      <tr>
-        <th class="d-w25p" scope="col">
-          Class
-        </th>
-        <th scope="col">
-          Output
-        </th>
-      </tr>
-    </thead>
-    <slot name="content" />
-  </table>
+  <div v-dt-scrollbar class="d-hmx464 d-bar8 d-ba d-bc-subtle">
+    <div>
+      <table class="d-table dialtone-doc-table d-fc-primary">
+        <thead class="d-bgc-primary d-ps-sticky d-zi-base1 d-t0">
+          <tr>
+            <th class="d-w25p d-p0 d-bbw0" scope="col">
+              <div class="d-p16 d-bb d-bc-default d-bbw1">
+                Class
+              </div>
+            </th>
+            <th class="d-p0 d-bbw0" scope="col">
+              <div class="d-p16 d-bb d-bc-default d-bbw1">
+                Output
+              </div>
+            </th>
+          </tr>
+        </thead>
+        <slot name="content" />
+      </table>
+    </div>
+  </div>
 </template>
 
 <script>