+++ apps/dialtone-documentation/docs/.vuepress/theme/components/MobileSidebar.vue
@@ -4,8 +4,8 @@
       class="
         d-px16
         d-ps-fixed d-w100p d-bgc-secondary d-h64 d-x0
-        lg:d-d-flex d-bb d-bc-default d-ai-center
-        d-jc-space-between d-t64 d-d-none
+        d-d-flex d-bb d-bc-default d-ai-center
+        d-jc-space-between d-t64 lg:d-d-none
       "
     >
       <dt-breadcrumbs :breadcrumbs="breadcrumbs" />