+++ apps/dialtone-documentation/docs/.vuepress/theme/components/Page.vue
@@ -3,7 +3,7 @@
     class="d-d-grid d-jc-center"
     :class="gridClass"
   >
-    <div class="d-p24 lg:d-pr24 lg:d-pt64">
+    <div class="d-p24 d-pt64 lg:d-pt24">
       <page-header />
       <content />
       <nav