+++ packages/dialtone-vue3/recipes/conversation_view/editor/editor.vue
@@ -20,19 +20,23 @@
       >
         <dt-tooltip
           v-for="button in buttonGroup.buttonGroup"
-          :key="`${buttonGroup.key}-${JSON.stringify(button.selector)}`"
+          :key="getButtonKey(buttonGroup.key, button.selector)"
           :message="button.tooltipMessage"
           placement="top"
         >
           <template #anchor>
             <dt-button
+              :ref="(el) => { buttonRefMap[getButtonRef(buttonGroup.key, button.selector)] = el }"
               :active="$refs.richTextEditor?.editor?.isActive(button.selector)"
               :aria-label="button.tooltipMessage"
               :data-qa="button.dataQA"
+              :tabindex="canFocus(getButtonRef(buttonGroup.key, button.selector)) ? 0 : -1"
               importance="clear"
               kind="muted"
               size="xs"
               @click="button.onClick()"
+              @keydown.right.stop="shiftActionBarFocusRight"
+              @keydown.left.stop="shiftActionBarFocusLeft"
             >
               <template #icon>
                 <component
@@ -71,13 +75,17 @@
             >
               <template #anchor>
                 <dt-button
+                  :ref="(el) => { buttonRefMap[getButtonRef('custom', 'link')] = el }"
                   :active="$refs.richTextEditor?.editor?.isActive(linkButton.selector)"
                   :aria-label="linkButton.tooltipMessage"
                   :data-qa="linkButton.dataQA"
+                  :tabindex="canFocus(getButtonRef('custom', 'link')) ? 0 : -1"
                   importance="clear"
                   kind="muted"
                   size="xs"
                   @click="linkButton.onClick()"
+                  @keydown.right.stop="shiftActionBarFocusRight"
+                  @keydown.left.stop="shiftActionBarFocusLeft"
                 >
                   <template #icon>
                     <component
@@ -206,7 +214,7 @@ import {
   DtIconCodeBlock,
   DtIconImage,
   DtIconItalic,
-  DtIconLightningBolt,
+  DtIconQuickReply,
   DtIconLink2,
   DtIconListBullet,
   DtIconListOrdered,
@@ -214,6 +222,7 @@ import {
   DtIconStrikethrough,
   DtIconUnderline,
 } from '@dialpad/dialtone-icons/vue3';
+import { ref } from 'vue';
 
 export default {
   compatConfig: { MODE: 3 },
@@ -226,7 +235,7 @@ export default {
     DtStack,
     DtInput,
     DtTooltip,
-    DtIconLightningBolt,
+    DtIconQuickReply,
     DtIconBold,
     DtIconItalic,
     DtIconUnderline,
@@ -549,6 +558,8 @@ export default {
 
       showLinkInput: false,
       linkInput: '',
+      currentButtonRefIndex: 0,
+      buttonRefMap: ref({}),
     };
   },
 
@@ -574,6 +585,17 @@ export default {
       return this.showListItemsButton || this.showOrderedListButton;
     },
 
+    orderedRefs () {
+      const refs = this.buttonGroups.reduce(function (acc, buttonData) {
+        buttonData.buttonGroup.forEach(button => {
+          acc.push(this.getButtonRef(buttonData.key, button.selector));
+        }, this);
+        return acc;
+      }.bind(this), []);
+      refs.push(this.getButtonRef('custom', 'link'));
+      return refs;
+    },
+
     buttonGroups () {
       const individualButtonStacks = this.individualButtons.map(buttonData => ({
         key: buttonData.selector,
@@ -594,7 +616,7 @@ export default {
           showBtn: this.showQuickRepliesButton,
           label: 'Quick reply',
           selector: 'quickReplies',
-          icon: DtIconLightningBolt,
+          icon: DtIconQuickReply,
           dataQA: 'dt-recipe-editor-quick-replies-btn',
           tooltipMessage: 'Quick Reply',
           onClick: this.onQuickRepliesClick,
@@ -737,6 +759,7 @@ export default {
         onClick: this.openLinkInput,
       };
     },
+
   },
 
   watch: {
@@ -901,6 +924,39 @@ export default {
       this.$emit('update:modelValue', event);
     },
 
+    getButtonKey (key, selector) {
+      return `${key}-${JSON.stringify(selector)}`;
+    },
+
+    // Unique Button Ref Key to identify ref
+    getButtonRef (key, selector) {
+      return `${this.getButtonKey(key, selector)}-ref`;
+    },
+
+    /**
+     * Determines if an element in the action bar button list is focusable with tab key
+     * @param {string} refKey - unique identifier for the ref element in DOM
+     */
+    canFocus (refKey) {
+      return refKey === this.orderedRefs[this.currentButtonRefIndex];
+    },
+
+    shiftActionBarFocusRight () {
+      this.shiftButtonRefIndex(1);
+    },
+
+    shiftActionBarFocusLeft () {
+      this.shiftButtonRefIndex(-1);
+    },
+
+    shiftButtonRefIndex (shiftAmount) {
+      const previousRef = this.buttonRefMap[this.orderedRefs[this.currentButtonRefIndex]];
+      const index = (this.currentButtonRefIndex + shiftAmount) % this.orderedRefs.length;
+      this.currentButtonRefIndex = index >= 0 ? index : this.orderedRefs.length + index;
+      const currentRef = this.buttonRefMap[this.orderedRefs[this.currentButtonRefIndex]];
+      previousRef.$el.blur();
+      currentRef.$el.focus();
+    },
   },
 };
 </script>